Extracted credentials generation out from jacktrip-init into a new jacktrip-credentials service, which now runs before bthelper@hci0. This ensures that credentials are generated before the BLE beacon is activated.
Fixed an issue with avahi where it was creating new tmp bind mounts
every time jacktrip-init was restarted. Also eliminated a systemctl
restart command inside jacktrip-init.sh by making the avahi-daemon
service always run after jacktrip-init.

Updated audio interface detection in jacktrip-init so that if no
interface is detected, it will repeatedly retry every 5 seconds.
This allows one to plug in a USB interface after powering on the
device, and still have everything startup and work as expected.
When an interface is not detected, it will also now turn the green
and red LED lights on and off every 5 seconds to provide a visual
indication of the problem.

Note that earlier changes modified the light behavior so that red
solid on and green off means it is attempting to connect to the
Internet (if it stays that way for a long time, it indicates a
network failure). After an interface is successfully detected,
jacktrip-init will now turn green and red leds to solid on to
signal a successful startup of jacktrip-agent.